+++ src/main/java/kr/KWGraduate/BookPharmacy/dto/board/response/BoardMyPageDto.java
@@ -1,7 +1,7 @@
 package kr.KWGraduate.BookPharmacy.dto.board.response;
 
 import kr.KWGraduate.BookPharmacy.entity.Board;
-import kr.KWGraduate.BookPharmacy.enums.Keyword;
+import kr.KWGraduate.BookPharmacy.entity.Keyword;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
