+++ src/main/java/kr/KWGraduate/BookPharmacy/entity/Board.java
@@ -1,13 +1,8 @@
 package kr.KWGraduate.BookPharmacy.entity;
 
 import jakarta.persistence.*;
-import kr.KWGraduate.BookPharmacy.dto.question.QuesAndDis;
-import kr.KWGraduate.BookPharmacy.enums.Keyword;
 import kr.KWGraduate.BookPharmacy.enums.Status;
 import lombok.*;
-
-import java.time.LocalDate;
-import java.util.List;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -58,7 +53,4 @@ public class Board extends BaseTimeEntity{
         this.status = Status.PRESCRIBING;
     }
 
-    public List<QuesAndDis> getQuesAndDis(){
-        return keyword.getQuesAndDis();
-    }
 }
